requirements.txt
gevent_socketio==0.3.6
numpy==1.25.2
opencv_python==4.8.0.76
pynput==1.7.6
python-dotenv==1.0.0